add scrcpy

+++ src/linktools/ios/ios.py
@@ -244,27 +244,38 @@ class GoIOSDevice(BaseDevice):
         import paramiko
         from linktools.ssh import SSHClient
 
-        forward = self.forward(
-            local_port=utils.pick_unused_port(range(20000, 30000)),
-            remote_port=port,
-        )
+        forward = None
+        client = None
+        try:
+            forward = self.forward(
+                local_port=utils.get_free_port(),
+                remote_port=port,
+            )
 
-        class Client(SSHClient):
+            class Client(SSHClient):
 
-            def close(self):
-                super().close()
+                def close(self):
+                    try:
+                        super().close()
+                    finally:
+                        forward.stop()
+
+            client = Client()
+            client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+            client.connect_with_pwd(
+                "localhost",
+                port=forward.local_port,
+                username=username,
+                password=password,
+            )
+        except:
+            if client is not None:
+                utils.ignore_error(client.close)
+            elif forward is not None:
                 forward.stop()
+            raise
 
-        ssh_client = Client()
-        ssh_client.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-        ssh_client.connect_with_pwd(
-            "localhost",
-            port=forward.local_port,
-            username=username,
-            password=password,
-        )
-
-        return ssh_client
+        return client
 
     def __repr__(self):
         return f"IOSDevice<{self.id}>"
@@ -277,17 +288,18 @@ class Forward(Stoppable):
     def __init__(self, ios: GoIOSDevice, local_port: int, remote_port: int):
         self._local_port = local_port
         self._remote_port = remote_port
-        self._process = ios.popen(
-            "forward",
-            local_port,
-            remote_port,
-            text=True,
-            bufsize=1,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
+        self._process = None
 
-        try:
+        def start():
+            self._process = ios.popen(
+                "forward",
+                local_port,
+                remote_port,
+                text=True,
+                bufsize=1,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+            )
             for out, err in self._process.fetch(timeout=10):
                 for line in (out or "").splitlines() + (err or "").splitlines():
                     data = utils.ignore_error(json.loads, args=(line,), default=None)
@@ -301,28 +313,26 @@ class Forward(Stoppable):
                             _logger.debug(f"Capture ios forward process output: {data['msg']}")
                             return
             raise GoIOSError("Run ios forward failed")
-        except:
-            self.stop()
-            raise
+
+        self._stop_on_error(start)
 
     def stop(self):
-        try:
-            _logger.debug(f"Kill ios proxy process")
-            self._process.kill()
-            self._process.wait(5)
-        except TimeoutExpired:
-            _logger.error(f"Proxy process did not finish normally")
+        if self._process is not None:
+            try:
+                _logger.debug(f"Kill ios proxy process")
+                self._process.kill()
+                self._process.wait(5)
+            except TimeoutExpired:
+                _logger.error(f"Proxy process did not finish normally")
 
 
 if __name__ == '__main__':
     import logging
+    from .. import rich
 
-    logging.basicConfig(level=logging.DEBUG)
+    rich.init_logging(level=logging.DEBUG, show_level=True)
+
     device = GoIOSDevice()
-    # # print(device.info)
-    # print(device.name)
-    # print(device.version)
-    # print(GoIOSDevice("11").info)
     # print(device.exec("ps", log_output=True))
     with device.ssh(22, "root", "alpine") as ssh:
         ssh.open_shell()
